+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Changed
 
+-   [band] props for `BarsLabels` to match `DendrogramLeafLabels`
 -   [all] added conditional exports to `package.json` files, specifying entry points for 'import' and 'require'
 -   [all] updated dependencies in `package.json` files to enable recent
     version of framer-motion